refactor/gossipsub: Implemented Deref/DerefMut and remove GossupsubStream::peers
This refactor would dereference to `Gossipsub` and expose its functions directly while also removing `peers` and use gossipsub internals to keep track of the peers that subscribe or unsubscribe. This would also fix the "leak" that is due to peers not being removed from the map when unsubscribing since the event itself, although it is a bit redundant to have this field when gossipsub also tracks this information.